+++ IG1App/GlassParapet.h
@@ -4,7 +4,7 @@
 class GlassParapet : public EntityWithTexture
 {
 public:
-	GlassParapet(GLdouble length,const char* texture, GLint alpha);
+	GlassParapet(GLdouble length,const char* texture, GLubyte alpha);
 	~GlassParapet() {};
 	void render(const glm::mat4& modelViewMat) const override;
 };
